# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Transfer Service API
  version: 0.1.0
  description: |
    Contract is the source of truth. Changes require human approval.
    Money amounts are integers in minor units (satang for THB). Never floats.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
tags:
  - name: health
    description: Probes for the platform
  - name: accounts
    description: Accounts owned by the authenticated customer
  - name: transfers
    description: Idempotent intra-bank money movement
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Liveness probe
      tags: [health]
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, example: ok }
  /v1/accounts/{accountId}:
    get:
      operationId: getAccount
      summary: Get an account owned by the caller
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Account
          headers:
            X-Trace-Id: { $ref: "#/components/headers/TraceId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/transfers:
    post:
      operationId: createTransfer
      summary: Transfer money between two accounts in the bank (idempotent)
      tags: [transfers]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Unique per customer. Retrying with the same key and body returns the original transfer.
          schema: { type: string, minLength: 1, maxLength: 64, pattern: "^[A-Za-z0-9_-]+$" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTransferRequest" }
      responses:
        "201":
          description: Transfer created
          headers:
            X-Trace-Id: { $ref: "#/components/headers/TraceId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transfer" }
        "200":
          description: Replay of an earlier request with the same Idempotency-Key and body
          headers:
            X-Trace-Id: { $ref: "#/components/headers/TraceId" }
            Idempotent-Replayed:
              description: Always "true" on a replay
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transfer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/BusinessRuleViolation" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/transfers/{transferId}:
    get:
      operationId: getTransfer
      summary: Get a transfer requested by the caller
      tags: [transfers]
      parameters:
        - name: transferId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Transfer
          headers:
            X-Trace-Id: { $ref: "#/components/headers/TraceId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transfer" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string, format: uuid }
  headers:
    TraceId:
      description: Correlation id; echoes the request's X-Trace-Id or a generated one
      schema: { type: string }
  responses:
    BadRequest:
      description: "VALIDATION_ERROR"
      headers:
        X-Trace-Id: { $ref: "#/components/headers/TraceId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: "UNAUTHORIZED"
      headers:
        X-Trace-Id: { $ref: "#/components/headers/TraceId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: "ACCOUNT_NOT_FOUND | TRANSFER_NOT_FOUND (also returned for resources the caller does not own)"
      headers:
        X-Trace-Id: { $ref: "#/components/headers/TraceId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BusinessRuleViolation:
      description: "INSUFFICIENT_FUNDS | ACCOUNT_INACTIVE | CURRENCY_MISMATCH | IDEMPOTENCY_KEY_REUSED"
      headers:
        X-Trace-Id: { $ref: "#/components/headers/TraceId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: "INTERNAL_ERROR (no internal details are exposed)"
      headers:
        X-Trace-Id: { $ref: "#/components/headers/TraceId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Money:
      type: object
      required: [amount, currency]
      additionalProperties: false
      properties:
        amount:
          type: integer
          format: int64
          description: Minor units (satang). 150000 = 1,500.00 THB
          example: 150000
        currency:
          type: string
          enum: [THB]
    Account:
      type: object
      required: [id, balance, status]
      properties:
        id: { type: string, format: uuid }
        balance: { $ref: "#/components/schemas/Money" }
        status: { type: string, enum: [ACTIVE, FROZEN, CLOSED] }
    CreateTransferRequest:
      type: object
      required: [fromAccountId, toAccountId, amount]
      additionalProperties: false
      properties:
        fromAccountId: { type: string, format: uuid }
        toAccountId: { type: string, format: uuid }
        amount:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: amount.amount must be > 0
        reference: { type: string, maxLength: 140 }
    Transfer:
      type: object
      required: [id, fromAccountId, toAccountId, amount, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        fromAccountId: { type: string, format: uuid }
        toAccountId: { type: string, format: uuid }
        amount: { $ref: "#/components/schemas/Money" }
        reference: { type: string }
        status: { type: string, enum: [COMPLETED] }
        createdAt: { type: string, format: date-time }
    Error:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - NOT_FOUND
            - ACCOUNT_NOT_FOUND
            - TRANSFER_NOT_FOUND
            - INSUFFICIENT_FUNDS
            - ACCOUNT_INACTIVE
            - CURRENCY_MISMATCH
            - IDEMPOTENCY_KEY_REUSED
            - INTERNAL_ERROR
        message: { type: string }
        traceId: { type: string }